refactor(queue): extract AI-slop-advisory gating/orchestration into its own module
Part of #4013's module-split sequence (step 4, after transient-locks.ts
and signal-snapshot.ts): shouldCollectSlopEvidence, shouldRunSlopAiAdvisory,
and runAiSlopForAdvisory move to src/queue/slop-detection.ts. Pure
mechanical move, no behavior change -- a re-export shim keeps this
file's own internal callers and existing test imports working
unchanged.

buildAiReviewDiff and buildSecretScanDiff move to src/review/review-diff.ts
(their natural existing home -- both already wrapped
buildUnifiedReviewDiff there) rather than staying in processors.ts,
which would have made the new slop-detection.ts circularly import this
file.

Closed a genuine, previously-invisible coverage gap surfaced by
isolating this code in its own small file: the BYOK
declared-provider-matches-stored-key branch had no test exercising it
either way, so two tests were added covering both the match and
mismatch outcomes.
